+++ paragami/psdmatrix_patterns.py
@@ -318,7 +318,7 @@ class PSDSymmetricMatrixPattern(Pattern):
     def flat_indices(self, folded_bool, free=None):
         # If no indices are specified, save time and return an empty array.
         if not np.any(folded_bool):
-            return np.array([])
+            return np.array([], dtype=int)
 
         free = self._free_with_default(free)
         shape_ok, err_msg = self._validate_folded_shape(folded_bool)
@@ -326,7 +326,7 @@ class PSDSymmetricMatrixPattern(Pattern):
             raise ValueError(err_msg)
         if not free:
             folded_indices = self.fold(
-                np.arange(self.flat_length(False)),
+                np.arange(self.flat_length(False), dtype=int),
                 validate_value=False, free=False)
             return folded_indices[folded_bool]
         else:
@@ -335,7 +335,7 @@ class PSDSymmetricMatrixPattern(Pattern):
             # pattern may be complicated and will
             # probably not make much of a difference in practice.
             if np.any(folded_bool):
-                return np.arange(self.flat_length(True))
+                return np.arange(self.flat_length(True), dtype=int)
             else:
                 return np.array([])
