+++ paragami/simplex_patterns.py
@@ -182,7 +182,7 @@ class SimplexArrayPattern(Pattern):
     def flat_indices(self, folded_bool, free=None):
         # If no indices are specified, save time and return an empty array.
         if not np.any(folded_bool):
-            return np.array([])
+            return np.array([], dtype=int)
 
         free = self._free_with_default(free)
         shape_ok, err_msg = self._validate_folded_shape(folded_bool)
@@ -190,7 +190,7 @@ class SimplexArrayPattern(Pattern):
             raise ValueError(err_msg)
         if not free:
             folded_indices = self.fold(
-                np.arange(self.flat_length(False)),
+                np.arange(self.flat_length(False), dtype=int),
                 validate_value=False, free=False)
             return folded_indices[folded_bool]
         else:
@@ -206,7 +206,8 @@ class SimplexArrayPattern(Pattern):
                 if np.any(folded_bool[ind]):
                     free_inds = np.arange(
                         offset * free_simplex_length,
-                        (offset + 1) * free_simplex_length)
+                        (offset + 1) * free_simplex_length,
+                        dtype=int)
                     indices.append(free_inds)
                 offset += 1
             if len(indices) > 0:
